Find number of Subscriptions
A new OTT streaming service was recently started in Bengaluru.A group of N friends in Bnegaluru want to buy OTT subscriptions. We know that 6 people can share one OTT subscription. Also, the cost of one OTT subscription is X rupees. Determine the minimum total cost that the group of N friends will incur so that everyone in the group is able to use OTT.

Sample Input:
12 250
Sample Output:
500
Explanation
There are 12 people in the group. Therefore they will have to buy 2 subscriptions. Therefore the total cost incurred is 500.