Discount Benefits
There are N items in a shop. You know that the price of the i-th item is Ai​​. Boy wants to buy all the N items. There is also a discount coupon that costs X rupees and reduces the cost of every item by Y rupees. If the price of an item was initially ≤Y, it becomes free, i.e, costs 0.
Determine whether boy should buy the discount coupon or not. Boy will buy the discount coupon if and only if the total price he pays after buying the discount coupon is strictly less than the price he pays without buying the discount coupon.

Input
The first line contains 3 space-separated integers - N, X and Y.
The second line contains N space separated integers A1, A2, ........., An.

Constraints
1 ≤ N ≤ 100
1 ≤ Ai, X, Y ≤ 105
Output
Return "COUPON" if boy will buy the discount coupon otherwise return "NO COUPON".

Example
Sample Input
4 30 10
15 8 22 6
Sample Output
COUPON
Explanation
The original cost of the items is 15+8+22+6=51.
Buying the coupon costs 30, and after buying it the cost of buying all the items is 5+0+12+0=17.
The total cost of buying everything with the coupon is 30+17=47, , which is strictly less than 51. So, Boy will buy the coupon.