Find Better Deal
There are 2 stores in Switzerland and both sell the same product. The first store sells the product for 100 rupees whereas the second store sells it for 200 rupees.
It is the holiday season and both stores have announced a special discount. The first store is providing a discount of A percent on its product and the second store is providing a discount of B percent on its product.
Gaurav is wondering which store is selling the product at a cheaper price after the discount has been applied. Can you help him identify the better deal?

Input
The first line of input will contain a single integer T, denoting the number of test cases.
Each test case consists of a single line of input containing two space-separated integers A and B denoting the discount provided by the first and second store respectively.

Constraints
1 ≤ T ≤ 1000
1 ≤ A, B ≤ 100
Output
For each test case, output First if the first store is cheaper, Second if the second store is cheaper, and Both if both the stores are selling the product for the same price after discount.
Example
Sample Input
2
5 20
10 60
Sample Output
First
Second
Explanation
Test case 1:  The first store has a discount of 5%. Thus, the final price of product at first store would be 95.
The second store has a discount of 20%. Thus, the final price of the product at the second store would be 160. The product at the first store would be cheaper.
Test case 2: The first store has a discount of 10%. Thus, the final price of product at first store would be 90.
The second store has a discount of 60%. Thus, the final price of the product at the second store would be 80. The product at the second store would be cheaper.