Single-use Attack
Ram is playing a video game, and is now fighting the final boss. The boss has H health points. Each attack of Ram reduces the health of the boss by X. Ram also has a special attack that can be used at most once, and will decrease the health of the boss by Y. Ram wins when the health of the boss is ≤ 0.
What is the minimum number of attacks needed by Ram to win?

Input
The first line of input will contain a single integer T, denoting the number of test cases.
The first and only line of each test case will contain three space-separated integers H, X, Y - the parameters described in the statement.

Constraints
1 ≤ T ≤ 1000
1 ≤ X < Y ≤ H ≤ 100
Output
For each test case, output on a new line the minimum number of attacks needed by Ram to win.
Example
Sample Input
2
100 25 40
100 29 45
Sample Output
4
3
Explanation
Test case 1: Ram can attack the boss 4 times normally. This results in 25 + 25 + 25 + 25 = 100 damage, which is enough to defeat the boss.
Test case 2: Ram can attack the boss 2 times normally, then use the special attack. This results in 29 + 29 + 45 = 103 damage, which is enough to defeat the boss.