Insurance
Sia bought car insurance. The policy of the insurance is:




The maximum rebatable amount for any damage is Rs X lakhs.

If the amount required for repairing the damage is ≤ X lakhs, that amount is rebated in full.



Sia's car meets an accident and required Rs Y lakhs for repairing. Determine the amount that will be rebated by the insurance company.

Input
The first line of input will contain a single integer T, denoting the number of test cases.
The first and only line of each test case contains two space-separated integers X and Y.

Constraints
1 ≤ T ≤ 1000
1 ≤ X, Y ≤ 30
Output
For each test case, output the amount (in lakhs) that will be rebated by the insurance company.
Example
Sample Input
2
5 3
5 8
Sample Output
3
5
Explanation
Test case 1: The damages require only Rs 3 lakh which is below the upper cap, so the entire Rs 3 lakh will be rebated.
Test case 2: The damages require Rs 8 lakh which is above the upper cap, so only Rs 5 lakh will be rebated.